- Material::AtomCount() now returns double; - Reimplemented in Geant4GM and RootGM to be consistent with AgddGM, which uses these values as input - In test/agddv6/Materials.agdd: replaced the element mass fractions with atom numbers - In AGDDWriter::WriteMaterial(): write atom count instead of mass fraction

+++ vgm/packages/RootGM/include/RootGM/materials/Material.h
@@ -87,7 +87,7 @@ namespace RootGM {
       virtual int             NofElements() const;
       virtual VGM::IElement*  Element(int iel) const;     
       virtual double          MassFraction(int iel) const;
-      virtual int             AtomCount(int iel) const;
+      virtual double          AtomCount(int iel) const;
 
     protected:  
       Material();
